many commands now allow shortnames, lots of lint fixes
lint had a lot to complain. Addressed valid stuff, disabled invalid
complaints. The python code still uses type annotation extremely rarely,
as the python release on door1 is

All commands dealing with a single user now allow shortnames, that is
you may shorten the name as long as its is unique. If the result is
disambiguous and you are in interactive mode, it will list the
matches and repeats asking for a new shortname until you have 1 or 0
results.
Interactive mode is when called without any command line parameters.
Batchmode in contrast is meant for unattended use in scripts and
therefore disallows user interaction.